fix: correct TuiStats size comment, revert cost_cents to u32
TuiStats is now 40 bytes instead of 48 bytes. The cost_cents does not
need widening - it wastes 8 bytes once padding is accounted for.
TuiStats only aggregates for Session/Message/Day.

u32 for cost_cents supports up to $43M per day, which is more than
sufficient (current usage: $177K over 3 months).

+++ src/contribution_cache/single_message.rs
@@ -122,7 +122,7 @@ impl PackedStatsDate {
             output_tokens: self.output_tokens() as u64,
             reasoning_tokens: self.reasoning_tokens() as u64,
             cached_tokens: self.cached_tokens() as u64,
-            cost_cents: self.cost_cents() as u64,
+            cost_cents: self.cost_cents() as u32,
             tool_calls: self.tool_calls() as u32,
         }
     }
